refactor: 편리한 테스트를 위해 도메인 간소화
