+++ src/main/java/io/github/daegwon/scheduled_transfer/domain/scheduled_transfer/ScheduledTransfer.java
@@ -1,9 +1,10 @@
 package io.github.daegwon.scheduled_transfer.domain.scheduled_transfer;
 
 
-import io.github.daegwon.scheduled_transfer.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -11,20 +12,17 @@ import java.time.LocalDateTime;
 @Getter
 @Entity
 @Table(name = "scheduled_transfers")
-public class ScheduledTransfer extends BaseEntity {
+public class ScheduledTransfer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "from_account_number", nullable = false)
-    private String fromAccountNumber;
+    @Column(nullable = false)
+    private String fromAccount;
 
-    @Column(name = "to_account_number", nullable = false)
-    private String toAccountNumber;
-
-    @Column(name = "to_bank_code", nullable = false)
-    private String toBankCode;
+    @Column(nullable = false)
+    private String toAccount;
 
     @Column(precision = 15, scale = 2, nullable = false)
     private BigDecimal amount;
@@ -35,4 +33,19 @@ public class ScheduledTransfer extends BaseEntity {
 
     @Column(name = "scheduled_at", nullable = false)
     private LocalDateTime scheduledAt;
+
+    @CreatedDate
+    @Column(name = "created_at", updatable = false, nullable = false)
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    @Column(name = "updated_at", nullable = false)
+    private LocalDateTime updatedAt;
+
+    public enum TransferStatus {
+        PENDING,
+        PROCESS,
+        COMPLETED,
+        FAILED
+    }
 }
