--- src/main/java/io/github/daegwon/scheduled_transfer/domain/scheduled_transfer/TransferStatus.java
@@ -1,7 +1,0 @@
-package io.github.daegwon.scheduled_transfer.domain.scheduled_transfer;
-
-public enum TransferStatus {
-    PENDING,
-    COMPLETED,
-    FAILED
-}
