--- src/main/java/io/github/daegwon/scheduled_transfer/domain/user/User.java
@@ -1,22 +1,0 @@
-package io.github.daegwon.scheduled_transfer.domain.user;
-
-
-import io.github.daegwon.scheduled_transfer.domain.BaseEntity;
-import jakarta.persistence.*;
-import lombok.Getter;
-
-@Getter
-@Entity
-@Table(name = "users")
-public class User extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String email;
-}
